feat: create basic structures
